+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Unstable
 * Funcionalidade: Adicionados tipos "Outro" e "Palestra" para Reserva.
 * Correção: Corrigido tipo de Reserva "Assembleia".
+* Funcionalidade: Adicionada exibição de reservas de acordo com o auditório selecionado.
 * Funcionalidade: Adicionada função de abrir modal de reserva com horário de início preenchido.
 
 # 0.0.2